requirements.txt
Django==5.2
djangorestframework
PyYAML
Markdown
Pygments
django-filter
django-guardian
psycopg[binary]
djangorestframework-simplejwt
